Revert "Change income quantity type to Float"
This reverts commit 33fa78eb8f7e209cc9bcd42b26fe1d9bd1553fc2.

+++ src/schema/modules/income.ts
@@ -6,7 +6,7 @@ export const typeDefs = `
   type Income {
     uuid: String!
     date: Int!
-    quantity: Float!
+    quantity: Int!
     value: Int!
     bought: Int
     sold: Int
@@ -18,7 +18,7 @@ export const typeDefs = `
 
   input IncomeInput {
     date: Int!
-    quantity: Float!
+    quantity: Int!
     value: Int!
     bought: Int
     sold: Int
